# packages/protocol/openapi/sidecar.openapi.yaml
openapi: 3.0.0
paths:
...
  /children/{id}:
    get:
      operationId: DatasetController_getChild
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a4
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
...
  /assets/{id}:
    get:
      operationId: DatasetController_getAsset
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a4
    delete:
      operationId: DatasetController_deleteAsset
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a4
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /assets/{id}/preview:
    get:
      operationId: DatasetController_getAssetPreview
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a4
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /assets/{id}/update:
    post:
      operationId: DatasetController_updateAsset
      parameters: []
      responses:
        "201":
          description: ""
      tags: *a4
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
...
  /storage/assets/{id}/sync:
    post:
      operationId: DatasetController_syncAssetToStorage
      parameters: []
      responses:
        "201":
          description: ""
      tags: *a4
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /storage/export-artifacts/{id}/sync:
    post:
      operationId: DatasetController_syncExportArtifactToStorage
      parameters: []
      responses:
        "201":
          description: ""
      tags: *a4
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
...
  /storage/export-artifacts/{id}/share:
    get:
      operationId: DatasetController_getExportArtifactShareMetadata
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a4
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
...
  /books/jobs/{id}:
    get:
      operationId: BooksController_getJob
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a5
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /books/jobs/{id}/preview:
    get:
      operationId: BooksController_preview
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a5
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /books/jobs/{id}/export/pdf:
    post:
      operationId: BooksController_exportPdf
      parameters: []
      responses:
        "201":
          description: ""
      tags: *a5
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /books/jobs/{id}/export/long-image:
    post:
      operationId: BooksController_exportLongImage
      parameters: []
      responses:
        "201":
          description: ""
      tags: *a5
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
...
  /api/web-companion/sessions/{sessionId}:
    get:
      operationId: WebCompanionController_getSessionSummary
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a6
    parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
...
  /api/web-companion/sessions/{sessionId}/items:
    post:
      operationId: WebCompanionController_createUploadItems
      parameters: []
      responses:
        "201":
          description: ""
      tags: *a6
    parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
  /api/web-companion/sessions/{sessionId}/items/{uploadItemId}/commit:
    put:
      operationId: WebCompanionController_commitUploadItem
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a6
    parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
      - name: uploadItemId
        in: path
        required: true
        schema:
          type: string
...
  /api/web-companion/share/{shareToken}/access:
    get:
      operationId: WebCompanionController_accessSharedContent
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a6
    parameters:
      - name: shareToken
        in: path
        required: true
        schema:
          type: string
  /api/web-companion/share/{shareToken}/assets:
    get:
      operationId: WebCompanionController_getSharedAssets
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a6
    parameters:
      - name: shareToken
        in: path
        required: true
        schema:
          type: string
  /api/web-companion/share/{shareToken}/book:
    get:
      operationId: WebCompanionController_getSharedBook
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a6
    parameters:
      - name: shareToken
        in: path
        required: true
        schema:
          type: string
...
  /api/web-companion/direct-upload/sessions/{sessionId}/config:
    get:
      operationId: DirectUploadController_getSessionConfig
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a7
    parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
...
info:
  title: KidMemory Sidecar API
  description: Local sidecar API for KidMemory desktop application
  version: 1.0.0
  contact: {}
...
servers: []
...